% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickRareCurve.R
\name{quickRareCurve}
\alias{quickRareCurve}
\title{Visualizing read quality indices per sample}
\usage{
quickRareCurve(
  x,
  step = 1,
  sample,
  xlab = "Sample Size",
  ylab = "Species",
  label = TRUE,
  col,
  lty,
  max.cores = T,
  nCores = 1,
  ...
)
}
\description{
Visualizing read quality indices per sample
}
