% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaticsQ.R
\name{staticsQ}
\alias{staticsQ}
\title{Calculate read quality indices}
\usage{
staticsQ(dir = x)
}
\arguments{
\item{dir}{: dir is a directory contained fastq files.}
}
\value{
column 1 : Maximum quality in a read. column 2 : Minimum Quality in a read. column 3 : Maximum expected error
}
\description{
Calculate read quality indices
}
