% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeSamplenames.R
\name{takeSamplenames1}
\alias{takeSamplenames1}
\title{Change the name shorter}
\usage{
takeSamplenames1(x, sep = "")
}
\arguments{
\item{x}{: x is a character vector of sample names.}

\item{sep}{: sep is a character to split the name.
#'}
}
\description{
Change the name shorter
}
